# [TechDebt] Clarify status-handling logic in BlockBasedTableBuilder::WriteRawBlock