**Context:** 
Inside `BlockBasedTableBuilder::WriteRawBlock`, there are multiple places that change local variables `io_s` and `s` while
depend on them. This PR attempts to clarify the relevant logics so that it's easier to read and add places of changing these local variables later (like https://github.com/facebook/rocksdb/pull/9342.) without changing the current behavior.

**Summary:**
- Shorten the lifetime of local var `io_s` and `s` as much as possible to avoid if-else branches by early return

**Test**
- Reasoned against original behavior to verify new changes do not break existing behaviors. 
- Rely on CI tests since we are not changing current behavior. 
- db_bench `./db_bench --benchmarks="fillseq"` X 5:
  - before the change: fillseq  :      3.6086 micros/op, 282474.6 ops/sec, 31.24 MB/s
  - after the change: fillseq  :       3.4078 micros/op 293929.2 ops/sec;   32.52 MB/s