@pdillinger 2022-01-18
Wow, the existing code is nightmarish. Your changes are small improvement.

* It looks like all callers to `SetIOStatus` also call `SetStatus` immediately after. Why not instead make `SetIOStatus` call `SetStatus`? I don't really know why we even have separate state for them. @anand1976 do you know?
* Why do we need `Status s` and `IOStatus io_s` variables that live for the whole function? Can they be temporaries with minimum lifetime, as in
```
{
  IOStatus io_s = r->file->Append(block_contents);
  if (!io_s.ok()) {
    SetIOStatus(io_s);
    // SetStatus would go here if not implied by SetIOStatus
    return;
  }
}
```

This sort of approach should greatly reduce the nesting depth (if not also cyclomatic complexity). The early return idiom might not be appropriate in cases where there's significant "clean up" not easily handled with RAII, but here it seems appropriate.