@hx235 2022-01-19
@pdillinger 
> Why do we need Status s and IOStatus io_s variables that live for the whole function? Can they be temporaries with minimum lifetime, as in

Good suggestion - fixed this, although there is 2 cases where we can't do early return cuz some existing behavior continues despite errors, which I don't why we did that but that's the current behavior.

> I don't really know why we even have separate state for them. @anand1976 do you know?
Looked into the history and found #6487.  Seems that IOStatus has more info than Status such as "such as error scope, error retryable attribute" and it was used from the bottom layer and passed upward into the write path (including block based table). That seems to be a legitimate reason to have two separate statuses and and two separate SetXX function. 

> It looks like all callers to SetIOStatus also call SetStatus immediately after. Why not instead make SetIOStatus call SetStatus?