@hx235 2022-01-19
Update:
- Adopted suggestion in shortening the lifetime of local variable io_s and s
- Left some PR comment and note responding to Peter's questions
- Ready for review again