@anand1976 2022-01-19
> > I don't really know why we even have separate state for them. @anand1976 do you know?
> 
> Looked into the history and found #6487. Seems like IOStatus has more info than Status "such as error scope, error retryable attribute" and it was used in the bottom IO layer and being passed upward into the write path (including block based table). That seems to be a legitimate reason to have two separate statuses and two separate SetXX function.

Right. And higher up in the write path, we always consider the IOStatus, but Status is ignored. So I guess non-IO failures are not considered critical and hence the need to keep them separate.

> 
> > It looks like all callers to SetIOStatus also call SetStatus immediately after. Why not instead make SetIOStatus call SetStatus?
> 
> I am not strongly against the idea of a nested call to SetStatus inside SetIOStatus, although I like them being independent so that there is no side-effect when calling SetIOStatus, unless we want to explicitly make this "side-effect" as a correct effect and document it as an invariant in the API cuz "SetIOStatus" does not sound obvious that it will SetStatus