@hx235 2022-01-20
@zhichao-cao I can indeed see this pattern in other places too. But the existing logic in WriteRawBlock by reusing the same s and io_s gives us many nested loop and hard to reason and add code on top of if.