@hx235 2022-01-20
@zhichao-cao I can indeed see this pattern in other places. But the existing logic in WriteRawBlock of reusing the same s and io_s gives us many nested loop and make it hard to reason/ add code on top of if.

What do you think of naming each io_s and s differently in WriteRawBlock:
`IOStatus io_s_append_trailer = r->file->Append(Slice(trailer.data(), trailer.size()));`