@zhichao-cao 2022-01-20
> What do you think of naming each io_s and s differently in my revision to `WriteRawBlock` and refrain from using same naming of `s` and `io_s`, like : `IOStatus io_s_append_trailer = r->file->Append(Slice(trailer.data(), trailer.size()));` `Status s_insert_block_in_cache = InsertBlockInCacheHelper(block_contents, handle, block_type, is_top_level_filter_block);`

Might be good for the "local" Status or IOStatus. Maybe shorter name, 5 words like "s_insert_block_in_cache" is too long.