@ajkr 2022-01-20
> I am not strongly against the idea of a nested call to SetStatus inside SetIOStatus, although I like them being independent so that there is no side-effect when calling SetIOStatus, unless we want to explicitly make this "side-effect" as a correct effect and document it as an invariant in the API cuz "SetIOStatus" does not sound obvious that it will SetStatus

The API doc for the getters suggest non-OK IOStatus implies non-OK Status:

```
  // Return non-ok iff some error has been detected.
  Status status() const override;                                                                                                                                                                                                                                           
                               
  // Return non-ok iff some error happens during IO.
  IOStatus io_status() const override;                                                                                                                                                                                                                                      
```

> Right. And higher up in the write path, we always consider the IOStatus, but Status is ignored. So I guess non-IO failures are not considered critical and hence the need to keep them separate.

@anand1976 To me this justified IOStatus but not Status. Hypothetical question: if IOStatus were named something non-IO specific, would we still need two?