@zhichao-cao 2022-01-20
> TODO:
> 
> * Investigate the rest of the two non-early return
> * Run db bench for performance since the change is in the write path.
> * Rest of the comments

db_bench might not be enough, agree with @ajkr , this code path is critical, any logical changes might result in a corner case failure.......