@hx235 2022-01-24
Update:
- Early-returned the rest of two cases, quoted @akankshamahajan15 to review the cache warming case
- Set Status in SetIOStatus and added comment

As for any performance concern on the very hot write path WriteRawBlock, is there any test I could do other than db_bench @ajkr @pdillinger? By the way, this PR is considered non-blocking to my #9342 but it's easier to read #9342's change to WriteRawBlock under this refactoring.