@pdillinger 2022-01-25
> As for any performance concern on the very hot write path WriteRawBlock, is there any test I could do other than db_bench @ajkr @pdillinger?

You can run a db_bench fillseq test to be safe, but the changes are mostly cosmetic. The set of code paths is nearly unchanged.